Station/requirements.txt
PyQt5
numpy
matplotlib
pyserial